+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ## 1.1.0 February 22, 2021
 
 - Enh #23: Add `SimpleEventDispatcher::clearEvents()` that clear all events in event dispatcher (vjik)
+- Enh #24: Add `SimpleEventDispatcher::getClassesEvents()` that return classes of dispatched events (vjik)
 
 ## 1.0.0 December 24, 2020